backend/db/requirements.txt
pymupdf
langchain
langchain-community
tqdm
chromadb
argparse
python-dotenv